refactor!: move ExportConfiguration to `ts` module
